feat/refactor(#212): modified cdn config structure; package restructured
