+++ playground-spring-boot-autoconfigure/src/main/java/com/oembedler/moon/playground/boot/PlaygroundPropertiesConfiguration.java
@@ -1,5 +1,6 @@
 package com.oembedler.moon.playground.boot;
 
+import com.oembedler.moon.playground.boot.properties.PlaygroundProperties;
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.validation.annotation.Validated;
