+++ playground-spring-boot-autoconfigure/src/main/java/com/oembedler/moon/playground/boot/properties/PlaygroundCdn.java
@@ -1,0 +1,13 @@
+package com.oembedler.moon.playground.boot.properties;
+
+import lombok.Data;
+
+import javax.validation.constraints.NotEmpty;
+
+@Data
+public class PlaygroundCdn {
+
+    private boolean enabled;
+    @NotEmpty
+    private String version = "latest";
+}
